+++ .github/workflows/company_ci.yml
@@ -4,7 +4,7 @@ on:
   push:
     paths:
       - 'company/**'
-      - 'shared/**'
+      - 'packages/**'
 
 env:
   YARN_CACHE_FOLDER: .cache/yarn
