+++ README.md
@@ -17,7 +17,7 @@ Node は以下のバージョンを想定する。
 
 - company (企業側サイト)
 - customer (顧客側サイト)
-- shared (共通で使用する機能などまとめる)
+- packages (共通で使用する機能などまとめる)
 
 ## VSCode の設定
