+++ company/pages/index.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { formatNumber } from '@shared/utils'
+import { formatNumber } from '@packs/utils'
 import type { NextPage } from 'next'
 import Container from '@mui/material/Container'
 import Typography from '@mui/material/Typography'
