+++ customer/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../shared/tsconfig.base.json",
+  "extends": "../packages/tsconfig.base.json",
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
