+++ packages/components-next/.storybook/main.js
@@ -1,0 +1,12 @@
+module.exports = {
+  stories: ['../**/*.stories.@(js|jsx|ts|tsx)'],
+  "addons": [
+    "@storybook/addon-links",
+    "@storybook/addon-essentials",
+    "@storybook/addon-interactions"
+  ],
+  "framework": "@storybook/react",
+  "core": {
+    "builder": "@storybook/builder-webpack5"
+  }
+}
