+++ packages/components-next/components/input/InputRadioGroup/InputRadioGroup.tsx
@@ -1,9 +1,29 @@
 import { FC } from 'react'
-import { DIRECTION_TYPE } from '@shared/constants/other'
-import { InputRadioGroupProp } from './InputRadioGroupType'
+import { DirectionType, DIRECTION_TYPE } from '@packs/constants/other'
 import Radio from '@mui/material/Radio'
 import RadioGroup from '@mui/material/RadioGroup'
 import FormControlLabel from '@mui/material/FormControlLabel'
+
+type InputRadioGroupValue = number | string | boolean | null
+
+export type InputRadioGroupProp = {
+  /** 値 */
+  readonly modelValue: InputRadioGroupValue
+  /** チェックリスト */
+  readonly items: any[]
+  /** キーのプロパティ */
+  readonly itemValueKey: string
+  /** 値のプロパティ */
+  readonly itemLabelKey: string
+  /** 方向 */
+  readonly direction: DirectionType
+  /** 必須可否 */
+  readonly required: boolean
+  /** 活性フラグ */
+  readonly disabled: boolean
+  /** エラーメッセージ */
+  readonly errorMessages: string[]
+}
 
 export const InputRadioGroup: FC<InputRadioGroupProp> = (props) => {
   return (
