+++ packages/constants/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@shared/constants",
+  "name": "@packs/constants",
   "version": "1.0.0",
   "main": "index.ts",
   "license": "MIT"
