+++ packages/utils/lib/common/formatNumber.ts
