+++ packages/utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@shared/utils",
+  "name": "@packs/utils",
   "version": "1.0.0",
   "main": "index.ts",
   "license": "MIT",
