--- shared/package.json
@@ -1,7 +1,0 @@
-{
-  "name": "shared",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "devDependencies": {}
-}
